coderadar-graph/src/main/java/io/reflectoring/coderadar/graph/projectadministration/filepattern/adapter/ListFilePatternsOfProjectAdapter.java
package io.reflectoring.coderadar.graph.projectadministration.filepattern.adapter;

import io.reflectoring.coderadar.graph.projectadministration.filepattern.FilePatternMapper;
import io.reflectoring.coderadar.graph.projectadministration.filepattern.repository.FilePatternRepository;
import io.reflectoring.coderadar.projectadministration.domain.FilePattern;
import io.reflectoring.coderadar.projectadministration.port.driven.filepattern.ListFilePatternsOfProjectPort;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class ListFilePatternsOfProjectAdapter implements ListFilePatternsOfProjectPort {
  private final FilePatternRepository filePatternRepository;
  private final FilePatternMapper filePatternMapper = new FilePatternMapper();

  public ListFilePatternsOfProjectAdapter(FilePatternRepository filePatternRepository) {
    this.filePatternRepository = filePatternRepository;
  }

  @Override
  public List<FilePattern> listFilePatterns(long projectId) {
    return filePatternMapper.mapNodeEntities(filePatternRepository.findByProjectId(projectId));
  }
}
